requirements.txt
chainlit
pydantic-ai
tavily-python
httpx
